Separate main into packages
